Refactor init_optimizer_params
